nive

+++ Initial/Hard/hardening.sh
@@ -26,6 +26,7 @@ fi
 
 # Sed sshd_config
 sed_ssh() {
+    echo "\n---------- Doing SSH config ----------"
     sed -i.bak 's/.*\(#\)\?Port.*/Port 22/g' /etc/ssh/sshd_config
     sed -i.bak 's/.*\(#\)\?Protocol.*/Protocol 2/g' /etc/ssh/sshd_config
     sed -i.bak 's/.*\(#\)\?UsePrivilegeSeperation.*/UsePrivilegeSeperation yes/g' /etc/ssh/sshd_config
@@ -72,6 +73,7 @@ sed_ssh() {
 }
 
 fix_corrupt(){
+    echo "\n---------- Corrupt Packages ---------"
     if [ "$os_type" == "RHEL" ]; then
         echo "fixing corrupt packages"
         rpm -qf $(rpm -Va 2>&1 | grep -vE '^$|prelink:' | sed 's|.* /|/|') | sort -u
@@ -86,12 +88,13 @@ fix_corrupt(){
 }
 
 reset_environment() {
-    echo "\n Resetting PATH variable"
+
+    echo "\n---------- Resetting PATH variable ----------"
     echo "PATH=\"/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games\"" > /etc/environment
 }
 
 check_ssh_keys() {
-    echo "\n Checking for ssh keys..."
+    echo "\n---------- Checking for ssh keys... ----------"
     s="sudo"
     $s cat /root/ssh/sshd_config | grep -i AuthorizedKeysFile
     $s head -n 20 /home/*/.ssh/authorized_keys*
@@ -99,7 +102,8 @@ check_ssh_keys() {
 }
 
 kernel(){
-    echo "\n Resetting sysctl.conf file"
+    echo "\n---------- Doing stuff with Kernel ----------"
+    echo "Resetting sysctl.conf file"
     cat configs/sysctl.conf > /etc/sysctl.conf
     echo 0 > /proc/sys/kernel/unprivileged_userns_clone
 
@@ -132,7 +136,7 @@ kernel(){
 }
 
 aliases(){
-    echo "\n Resetting user bashrc files and profile file" 
+    echo "\n---------- Resetting user bashrc files and profile file ----------" 
 	for user in $(cat users.txt); do
         	cat configs/bashrc > /home/$user/.bashrc;
 	done;
@@ -141,12 +145,15 @@ aliases(){
 }
 
 configCmds(){
-	cat configs/adduser.conf > /etc/adduser.conf
-	cat configs/deluser.conf > /etc/deluser.conf
+    if [ "$os_type" == "Ubuntu" ]; then
+        echo "\n---------- setting addusers.conf and delusers.conf -----------"
+	    cat configs/adduser.conf > /etc/adduser.conf
+	    cat configs/deluser.conf > /etc/deluser.conf
+    fi
 }
 
 noIpv6(){
-    echo "/n Removing IPv6..."
+    echo "\n----------- Removing IPv6... -----------"
     interfaces=$(ifconfig | grep "flags" | awk '{print $1}' | tr -d ':')
     if [ os_type == "Debian" ]; then
         echo "removing IPv6 from sysctl"
@@ -203,6 +210,7 @@ noIpv6(){
 }
 
 cronConf(){
+    echo "\n----------- Cron Configs -----------"
     #works for both ubuntu and RHEL
     IFS=$'\n'
     cronOwnerConfs=$(find /etc -type f -o -type d -name 'cron*' -exec stat -c "%U %n" {} \;)
@@ -233,6 +241,7 @@ cronConf(){
 }
 
 accessModes(){
+    echo "\n---------- Setting Single User and Emergency Mode Security -----------"
     #single user mode
     singleFile="/usr/lib/systemd/system/rescue.service"
     singleSetting="ExecStart=-/usr/lib/systemd/systemd-sulogin-shell rescue"
@@ -268,10 +277,11 @@ accessModes(){
     #no blank/null passwords
     cp "/etc/pam.d/common-password" /tmp/common-password.bak
     sed -i  '/nullok/d' "/etc/pam.d/common-password"
-    echo "nullok removed from common-password"
+    echo "nullok removed from common-password, so no blank/null password are accepted"
 }
 
 maybeMalware(){
+    echo "\n----------- Trying to Find and Remove Malware -----------"
     REMOVE="john* netcat* iodine* kismet* medusa* hydra* fcrackzip* ayttm* empathy* nikto* logkeys* rdesktop* vinagre* openarena* openarena-server* minetest* minetest-server* ophcrack* crack* ldp* metasploit* wesnoth* freeciv* zenmap* knocker* bittorrent* torrent* p0f aircrack* aircrack-ng ettercap* irc* cl-irc* rsync* armagetron* postfix* nbtscan* cyphesis* endless-sky* hunt snmp* snmpd dsniff* lpd vino* netris* bestat* remmina netdiag inspircd* up.time uptimeagent chntpw* nfs* nfs-kernel-server* abc sqlmap acquisition bitcomet* bitlet* bitspirit* armitage airbase-ng* qbittorrent* ctorrent* ktorrent* rtorrent* deluge* tixati* frostwise vuse irssi transmission-gtk utorrent* exim4* crunch tomcat tomcat6 vncserver* tightvnc* tightvnc-common* tightvncserver* vnc4server* nmdb dhclient cryptcat* snort pryit gameconqueror* weplab lcrack dovecot* pop3 ember manaplus* xprobe* openra* ipscan* arp-scan* squid* heartbleeder* linuxdcpp* cmospwd* rfdump* cupp3* apparmor nis* ldap-utils prelink rsh-client rsh-redone-client* rsh-server quagga gssproxy iprutils sendmail nfs-utils ypserv tuned" 
     for package in $REMOVE; do
         if [ "$os_type" == "RHEL" ]; then
@@ -288,6 +298,7 @@ maybeMalware(){
 }
 
 deleteBad(){
+    echo "\n----------- Removing rhosts, host.equiv, etc. -----------"
     find / -name ".rhost" -exec rm -rf {} \;
     find / -name "host.equiv" -exec rm -rf {} \;
     find / -iname '*.xlsx' -delete
@@ -296,6 +307,7 @@ deleteBad(){
 }
 
 perms(){
+    echo "\n---------- Setting File Permissions -----------"
     chmod 755 /etc/resolvconf/resolv.conf.d/
     chmod 644 /etc/resolvconf/resolv.conf.d/base
     chmod 777 /etc/resolv.conf
@@ -427,6 +439,7 @@ perms(){
 }
 
 fstab(){
+    echo "\n---------- fstab configs -----------"
 	echo "tmpfs /run/shm tmpfs defaults,nodev,noexec,nosuid 0 0" >> /etc/fstab
 	echo "tmpfs /tmp tmpfs defaults,rw,nosuid,nodev,noexec,relatime 0 0" >> /etc/fstab
 	echo "tmpfs /var/tmp tmpfs defaults,nodev,noexec,nosuid 0 0" >> /etc/fstab
@@ -436,6 +449,7 @@ fstab(){
 }
 
 etcConf(){
+    echo "\n---------- etc configs ----------"
     echo "tty1" > /etc/securetty
 	echo "TMOUT=300" >> /etc/profile
 	echo "readonly TMOUT" >> /etc/profile
@@ -449,6 +463,7 @@ etcConf(){
 }
 
 dconfSettings(){
+    echo "\n---------- dconf Settings ----------"
     dconf reset -f /
 	gsettings set org.gnome.desktop.privacy remember-recent-files false
 	gsettings set org.gnome.desktop.media-handling automount false
@@ -458,6 +473,7 @@ dconfSettings(){
 }
 
 passwordPolicy(){
+    echo "\n---------- Setting Password Policy ----------"
     if [ "$os_type" == "RHEL" ]; then
         cp configs/password-auth.txt /etc/pam.d/password-auth
     elif [ "$os_type" == "Ubuntu" ]; then
@@ -467,6 +483,7 @@ passwordPolicy(){
 }
 
 other(){
+    echo "\n---------- Other Random Stuff To Run ----------"
     echo 0 > /proc/sys/kernel/unprivileged_userns_clone
     prelink -ua
     apt-get remove -y prelink
@@ -491,6 +508,7 @@ other(){
 }
 
 chattr(){
+    echo "\n---------- Setting Chattr ----------"
     chattr +ia /etc/passwd
     chattr +ia /etc/group
     chattr +ia /etc/shadow
